refactor due to future recognition tasks
