+++ src/detection/dataset_utils/copy_helper.py
