+++ src/detection/intersect_resolver.py
@@ -1,12 +1,12 @@
-from typing import Union, List
+from typing import List
 from abc import ABC, abstractmethod
 import torch
 
-import box_utils.shapes_util as su
-from box_utils.lines_util import find_bottom_left, find_bottom_right, find_top_right, find_top_left
+import detection.bounding_boxes.shapes as sh
+from detection.bounding_boxes.lines import find_bottom_left, find_bottom_right, find_top_right, find_top_left
 
 
-def find_intersecting_objects(boxes: List[su.Obb], threshold: float = 0.4) -> List[tuple[int, int]]:
+def find_intersecting_objects(boxes: List[sh.Obb], threshold: float = 0.4) -> List[tuple[int, int]]:
     """
     Finds interseceted boxes by IoU
     Args:
@@ -19,18 +19,18 @@ def find_intersecting_objects(boxes: List[su.Obb], threshold: float = 0.4) -> Li
 
     for i in range(len(boxes)):
         for j in range(i + 1, len(boxes)):
-            if su.IoU(boxes[i], boxes[j]) >= threshold:
+            if sh.IoU(boxes[i], boxes[j]) >= threshold:
                 intersecting_objects.append((i, j))
     
     return intersecting_objects
 
 
-def tensor_to_boxes(boxes: torch.Tensor) -> List[su.Obb]:
+def tensor_to_boxes(boxes: torch.Tensor) -> List[sh.Obb]:
     shapes = []
     for i in range(boxes.shape[0]):
         box = boxes[i]
         shapes.append(
-            su.yolo_result_to_obb(box)
+            sh.yolo_result_to_obb(box)
         )
             
     return shapes
@@ -39,28 +39,28 @@ def tensor_to_boxes(boxes: torch.Tensor) -> List[su.Obb]:
 class IntersectionResolver(ABC):
     @abstractmethod
     def __call__(self, 
-                 box_1: su.Obb, 
+                 box_1: sh.Obb, 
                  conf_1: float,
-                 box_2: su.Obb,
-                 conf_2: float) -> su.Obb:
+                 box_2: sh.Obb,
+                 conf_2: float) -> sh.Obb:
         pass
 
 
 class ByConfidenceResolver(IntersectionResolver):
     def __call__(self, 
-                 box_1: su.Obb, 
+                 box_1: sh.Obb, 
                  conf_1: float,
-                 box_2: su.Obb,
-                 conf_2: float) -> su.Obb:
+                 box_2: sh.Obb,
+                 conf_2: float) -> sh.Obb:
         return box_1 if conf_1 > conf_2 else box_2
 
 
 class UnionResolver(IntersectionResolver):
     def __call__(self, 
-                 box_1: su.Obb, 
+                 box_1: sh.Obb, 
                  conf_1: float,
-                 box_2: su.Obb,
-                 conf_2: float) -> su.Obb:
+                 box_2: sh.Obb,
+                 conf_2: float) -> sh.Obb:
         x_list = [box_1[i] for i in range(0, 8, 2)] + [box_2[i] for i in range(0, 8, 2)] 
         y_list = [box_1[i] for i in range(1, 8, 2)] + [box_2[i] for i in range(1, 8, 2)] 
 
@@ -69,7 +69,7 @@ class UnionResolver(IntersectionResolver):
         max_y = max(y_list)
         min_y = min(y_list)
 
-        return su.Obb(min_x, min_y, max_x, min_y, max_x, max_y, min_x, max_y)
+        return sh.Obb(min_x, min_y, max_x, min_y, max_x, max_y, min_x, max_y)
 
 
 def build_resolver_by_name(resolver_name: str | None) -> IntersectionResolver | None:
@@ -81,28 +81,28 @@ def build_resolver_by_name(resolver_name: str | None) -> IntersectionResolver | 
         resolver (IntersectionResolver | None)
     """
     if resolver_name == "ByConfResolver":
-        from intersect_resolver import ByConfidenceResolver
+        from detection.intersect_resolver import ByConfidenceResolver
         return ByConfidenceResolver()
     if resolver_name == "MeanResolver":
-        from intersect_resolver import MeanResolver
+        from detection.intersect_resolver import MeanResolver
         return MeanResolver()
     if resolver_name == "UnionResolver":
-        from intersect_resolver import UnionResolver
+        from detection.intersect_resolver import UnionResolver
         return UnionResolver()
     
 
 class MeanResolver(IntersectionResolver):    
     def __call__(self, 
-                 box_1: su.Obb, 
+                 box_1: sh.Obb, 
                  conf_1: float,
-                 box_2: su.Obb,
-                 conf_2: float) -> su.Obb:
-        box_1 = su.Obb(*find_top_left(box_1), 
+                 box_2: sh.Obb,
+                 conf_2: float) -> sh.Obb:
+        box_1 = sh.Obb(*find_top_left(box_1), 
                        *find_top_right(box_1), 
                        *find_bottom_right(box_1), 
                        *find_bottom_left(box_1))
 
-        box_2 = su.Obb(*find_top_left(box_2), 
+        box_2 = sh.Obb(*find_top_left(box_2), 
                        *find_top_right(box_2), 
                        *find_bottom_right(box_2), 
                        *find_bottom_left(box_2))
@@ -111,7 +111,7 @@ class MeanResolver(IntersectionResolver):
         for i in range(8):
             result_box[i] = (box_1[i] + box_2[i]) / 2
         
-        return su.Obb(*result_box)
+        return sh.Obb(*result_box)
 
 
 def resolve_intersected_objects(
@@ -119,7 +119,7 @@ def resolve_intersected_objects(
         confs: torch.Tensor,
         threshold: float, 
         resolver: IntersectionResolver
-    ) -> List[su.Obb]:
+    ) -> List[sh.Obb]:
     boxes = tensor_to_boxes(boxes)
     confs = confs.to("cpu")
     intersected_boxes = find_intersecting_objects(boxes, threshold)
